pyproject.toml
[project]
name = "drive-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
